+++ CHANGELOG.md
@@ -48,8 +48,11 @@ CLI command and its behaviour. There are no guarantees of stability for the
   - CommonJS (`.cjs`) (#632)
   - Qt .pro (`.pro`) (#632)
   - Textile (`.textile`) (#712)
+  - Svelte components (`.svelte`)
 - More files are recognised:
   - Clang format (`.clang-format`) (#632)
+  - Browserslist config (`.browserslist`)
+  - Prettier config (`.prettierrc`) and ignored files (`.prettierignore`)
 - Added loglevel argument to pytest and skip one test if loglevel is too high
   (#645).
 - `--add-license-concluded`, `--creator-person`, and `--creator-organization`